+++ packages/eyes-storybook/src/errMessages.js
@@ -33,10 +33,6 @@ ${chalk.green(
 
 `;
 
-const startStorybookFailMsg = chalk.red(
-  'Could not find a storybook URL to test. This might be either because there was an error when starting the storybook dev server, or the parameter "storybookUrl" was not specified. Specify a URL with the "-u" parameter, or with "storybookUrl" in the config file (by default the config file is located at <project-folder>/applitools.config.js>).\n',
-);
-
 function refineErrorMessage({prefix, error}) {
   return `${prefix} ${error.message.replace('Evaluation failed: ', '')}`;
 }
@@ -45,6 +41,5 @@ module.exports = {
   missingApiKeyFailMsg,
   missingAppNameAndPackageJsonFailMsg,
   missingAppNameInPackageJsonFailMsg,
-  startStorybookFailMsg,
   refineErrorMessage,
 };