+++ packages/eyes-storybook/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- add `startStorybookServerTimeout` config param
 
 ## 3.23.1 - 2021/8/23